Clock.now's return value should be opaque to core
Clock.now may return different values on different platforms, as we want
to support whatever makes sense on each platform. For example, for
browsers 'performance.now' will be used, which is milliseconds since
the "time origin". On another platform, it may make sense to store a
full timestamp, a value in nanoseconds, or some other number that
matches the conventions on that platform

Core shouldn't care what this returns exactly, just that it returns the
same thing consistently