Make sure the $browser global is always available
We need $browser to be available at all times so that we can use it for
things like the 'skip' annotations. This was problematic as it was
created in a 'before' hook, which run just before a scenario starts, so
if a skip annotation required checking the $browser object then we'd
crash

The $browser global needed to be in a before hook because it used the
Maze Runner URLs for generating URLs to fixtures and those were only
available after configuration is setup. This is now handled by a
separate $url_generator object, which means the $browser object can be
made available at all times